Add totalUser stats

+++ common/src/main/java/urlshortener/common/domain/Stats.java
@@ -6,14 +6,14 @@ public class Stats {
 
     private int upTime;
     private Long totalURL;
-    private int totalUser;
+    private Long totalUser;
     private int averageAccessURL;
     private int responseTime;
     private int memoryUsed;
     private int memoryAvailable;
     private List<String> topURL;
 
-    public Stats(int upTime, Long totalURL, int totalUser, int averageAccessURL, int responseTime,
+    public Stats(int upTime, Long totalURL, Long totalUser, int averageAccessURL, int responseTime,
                  int memoryUsed, int memoryAvailable, List<String> topURL) {
         this.upTime = upTime;
         this.totalURL = totalURL;
@@ -36,7 +36,7 @@ public class Stats {
         return totalURL;
     }
 
-    public int getTotalUser() {
+    public Long getTotalUser() {
         return totalUser;
     }
